[GELC] remove autoboxing
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@1379 f329d543-caf0-0310-9063-dda96c69346f

+++ src/main/java/com/xpn/xwiki/doc/XWikiDocument.java
@@ -2652,9 +2652,9 @@ public class XWikiDocument {
             if (name.startsWith(start)) {
                 int pos = name.indexOf("_", start.length() + 1);
                 String prefix = name.substring(0, pos);
-                int num = Integer.decode(prefix.substring(prefix.lastIndexOf("_") + 1));
-                if (!objectsNumberDone.contains(num)){
-                    objectsNumberDone.add(num);
+                int num = Integer.decode(prefix.substring(prefix.lastIndexOf("_") + 1)).intValue();
+                if (!objectsNumberDone.contains(Integer.valueOf(num))){
+                    objectsNumberDone.add(Integer.valueOf(num));
                     objects.add(addObjectFromRequest(className, num, context));
                 }
             }
